Copy files so user does not have to mount Git dir
Copy files into image instead of making them
mount the git source directory.

This makes it so that people can use this
image without having to download the source files.

To develope with this image, one would still need
to mount the source files, so their changes are reflected
while debugging.